prints the executor memory usage while handling OSError
